+++ src/Activator/EnvironmentActivator.php
@@ -56,9 +56,24 @@ class EnvironmentActivator implements FeatureActivatorInterface
         }
 
         if ($this->forceRequest === true && !array_key_exists($name, $this->variables)) {
-            return filter_var(getenv($name), FILTER_VALIDATE_BOOLEAN);
+            return $this->getEnv($name);
         }
 
-        return filter_var(getenv($this->variables[$name]), FILTER_VALIDATE_BOOLEAN);
+        return $this->getEnv($this->variables[$name]);
+    }
+
+    /**
+     * Get enviroment value by $_ENV or getenv()
+     *
+     * @param $name
+     *
+     * @return string|null
+     */
+    private function getEnv($name)
+    {
+        return filter_var(
+            array_key_exists($name, $_ENV) ? $_ENV[$name] : getenv($name),
+            FILTER_VALIDATE_BOOLEAN
+        );
     }
 }